requirements.txt
# Environment 
gunicorn
yapf
flake8
psycopg2

# Testing
faker
watchdog

# Development
flask-sqlalchemy
flask-login
flask-dropzone
flask-mail
flask-moment
flask-wtf
python-dotenv
pillow
bootstrap-flask
flask
flask-whooshee
flask-avatars
flask-jwt-extended
flask-restful
flask-migrate
